1.0.6 requested changes
